Minor refactor - move Cell-related stuff to new CellUtil in new 'util' package
Will also move the CloudCityUtil there too.

But for now, the 'findRegisteredCell' and 'getMeasurementsModel' have been extracted
from LoggingServiceExtensions into this new CellUtil. There might be more candidates
but they will be moved as they become needed.
