+++ app/src/main/java/cloudcity/LoggingServiceExtensions.java
@@ -21,6 +21,7 @@ import cloudcity.networking.models.MeasurementsModel;
 import cloudcity.networking.models.MobileSignalNetworkDataModel;
 import cloudcity.networking.models.NetworkDataModel;
 import cloudcity.networking.models.NetworkDataModelRequest;
+import cloudcity.util.CellUtil;
 import de.fraunhofer.fokus.OpenMobileNetworkToolkit.DataProvider.CellInformations.CellInformation;
 import de.fraunhofer.fokus.OpenMobileNetworkToolkit.DataProvider.CellInformations.GSMInformation;
 import de.fraunhofer.fokus.OpenMobileNetworkToolkit.DataProvider.CellInformations.LTEInformation;
@@ -158,7 +159,7 @@ public class LoggingServiceExtensions {
         List<CellInformation> cellsInfo = dp.getCellInformation();
         List<CellInformation> signalInfo = dp.getSignalStrengthInformation();
 
-        CellInformation currentCell = findRegisteredCell(cellsInfo);
+        CellInformation currentCell = CellUtil.findRegisteredCell(cellsInfo);
         CellInformation currentSignal = null;
         // After testing on a real phone, apparently there's only one SignalStrengthInformation
         // CellInformation member in the list, so lets just take the first one if it's there
@@ -170,7 +171,7 @@ public class LoggingServiceExtensions {
 
         // Lets initialize our MeasurementModel for sending from the registered cell model, then overwrite it's values
         // with what we found in the SignalInformation
-        MeasurementsModel modelForSending = getMeasurementsModel(category, currentCell);
+        MeasurementsModel modelForSending = CellUtil.getMeasurementsModel(category, currentCell);
         updateMeasurementModelByCell(modelForSending, currentSignal);
 
         Location location = GPSMonitor.getLastLocation();
@@ -190,24 +191,7 @@ public class LoggingServiceExtensions {
         return dataModel;
     }
 
-    /**
-     * Finds a registered cell information in a list of cell informations
-     * @param cellList
-     * @return the first registered cell in the list, or null if no registered cells were found
-     */
-    public static CellInformation findRegisteredCell(@NonNull List<CellInformation> cellList) {
-        CellInformation retVal = null;
 
-        for (CellInformation ci: cellList) {
-            if (!ci.isRegistered()) {
-                continue;
-            }
-
-            retVal = ci;
-        }
-
-        return retVal;
-    }
 
     private static @NonNull CellInfoModel getCellInfoModel(String category, CellInformation currentCell) {
         CellInfoModel cellInfoModel = new CellInfoModel();
@@ -242,34 +226,6 @@ public class LoggingServiceExtensions {
         }
 
         return cellInfoModel;
-    }
-
-    private static @NonNull MeasurementsModel getMeasurementsModel(String category, CellInformation currentCell) {
-        MeasurementsModel measurements = new MeasurementsModel();
-
-        if (Objects.equals(category, "NR")) {
-            // New safety
-            if (currentCell instanceof NRInformation) {
-                NRInformation nrCell = (NRInformation) currentCell;
-                measurements.setCsirsrp(nrCell.getCsirsrp());
-                measurements.setCsirsrq(nrCell.getCsirsrq());
-                measurements.setCsisinr(nrCell.getCsisinr());
-                measurements.setSsrsrp(nrCell.getSsrsrp());
-                measurements.setSsrsrq(nrCell.getSsrsrq());
-                measurements.setSssinr(nrCell.getSssinr());
-            }
-        } else if (Objects.equals(category, "LTE")) {
-            if (currentCell instanceof LTEInformation) {
-                LTEInformation lteCell = (LTEInformation) currentCell;
-                measurements.setRsrp(lteCell.getRsrp());
-                measurements.setRsrq(lteCell.getRsrq());
-                measurements.setRssnr(lteCell.getRssnr());
-            }
-        } else {
-            /* In 3G no measurement data available set dummy data. */
-            measurements.setDummy(1);
-        }
-        return measurements;
     }
 
     private static void updateMeasurementModelByCell(@NonNull MeasurementsModel measurements, @NonNull CellInformation cellForUpdating) {
